+++ src/pga.jl
@@ -13,10 +13,10 @@ include("common.jl")
 
 
 #Basis
-const e1 = Odd{Float64}(Quaternion(0, -1.0, 0, 0), Quaternion(0.0))
-const e2 = Odd{Float64}(Quaternion(0, 0, -1.0, 0),Quaternion(0.0))
-const e3 = Odd{Float64}(Quaternion(0, 0, 0, -1.0),Quaternion(0.0))
-const e0 = Odd{Float64}(Quaternion(0.0), Quaternion(1.0,0,0,0))
+const e1 = Odd{Int8}(Quaternion(0, -1.0, 0, 0), Quaternion(0.0))
+const e2 = Odd{Int8}(Quaternion(0, 0, -1.0, 0),Quaternion(0.0))
+const e3 = Odd{Int8}(Quaternion(0, 0, 0, -1.0),Quaternion(0.0))
+const e0 = Odd{Int8}(Quaternion(0.0), Quaternion(1.0,0,0,0))
 const I3 = e1*e2*e3
 
 basPGA = [e1,e2,e3, e0]
